Devide event handler function from main, README update
